v1.4 (#36)
* Exchange rates service test+refactor II

* test_currency_to_czk (#28)

* Exchange rates endpoint test (#29)

* Login+Auth tests (#30)

* Database service tests (#31)

* rename constants to UPPER_CASE; payment history test

* test getting exchange rates

* Token tests, main_driver refactor

+++ services/database_service.py
@@ -2,12 +2,16 @@
 import pathlib
 import yaml
 
-database_path = pathlib.Path(__file__).parent.parent / "database"
+DATABASE_PATH = pathlib.Path(__file__).parent.parent / "database"
+USERS_PATH = DATABASE_PATH / "users.yaml"
+BANK_ACCOUNTS_PATH = DATABASE_PATH / "bank_accounts.yaml"
+PAYMENTS_PATH = DATABASE_PATH / "payments.yaml"
+EXCHANGE_RATES_PATH = DATABASE_PATH / "exchange_rates.yaml"
 
 
 def get_user(username):
     """Vrátí uživatele s požadovaným username, pokud existuje."""
-    with open(database_path / "users.yaml", "r", encoding="utf8") as file:
+    with open(USERS_PATH, "r", encoding="utf8") as file:
         users = yaml.safe_load(file)
     for user in users:
         if user["username"] == username:
@@ -17,51 +21,55 @@ def get_user(username):
 
 def get_bank_accounts(username):
     """Vrátí seznam všech účtů vlastněných uživatelem s username."""
-    with open(database_path / "bank_accounts.yaml", "r", encoding="utf8") as file:
+    with open(BANK_ACCOUNTS_PATH, "r", encoding="utf8") as file:
         bank_accounts = yaml.safe_load(file)
     return list(filter(lambda bank_account: bank_account["owner"] == username, bank_accounts))
 
 
 def get_bank_account(username, currency):
     """Vrátí seznam účtů patřících uživateli s 'username' ve měně 'currency'. (Této podmínce by měl vyhovovat jen jeden)."""
-    with open(database_path / "bank_accounts.yaml", "r", encoding="utf8") as file:
+    with open(BANK_ACCOUNTS_PATH, "r", encoding="utf8") as file:
         bank_accounts = yaml.safe_load(file)
     return list(filter(lambda bank_account: bank_account["owner"] == username and bank_account["currency"] == currency, bank_accounts))
 
 
 def set_bank_account_balance(iban, balance):
     """Nastaví u účtu s konkrétním 'iban' nový zůstatek 'balance'."""
-    with open(database_path / "bank_accounts.yaml", "r", encoding="utf8") as file:
+    if not BANK_ACCOUNTS_PATH.exists():
+        return
+    with open(BANK_ACCOUNTS_PATH, "r", encoding="utf8") as file:
         bank_accounts = yaml.safe_load(file)
     for bank_account in bank_accounts:
         if bank_account["iban"] == iban:
             bank_account["balance"] = balance
             break
-    with open(database_path / "bank_accounts.yaml", "w", encoding="utf8") as file:
+    with open(BANK_ACCOUNTS_PATH, "w", encoding="utf8") as file:
         yaml.dump(bank_accounts, file)
 
 
 def log_payment(iban, amount, timestamp):
     """Zaloguje platbu."""
-    with open(database_path / "payments.yaml", "a", encoding="utf8") as file:
+    if not PAYMENTS_PATH.exists():
+        return
+    with open(PAYMENTS_PATH, "a", encoding="utf8") as file:
         file.write(f'- iban: "{iban}"\n  value: {amount}\n  timestamp: {timestamp}\n')
 
 
 def get_payment_history(iban):
     """Vrátí historii plateb pro účet s daným 'iban'."""
-    with open(database_path / "payments.yaml", "r", encoding="utf8") as file:
+    with open(PAYMENTS_PATH, "r", encoding="utf8") as file:
         payments = yaml.safe_load(file)
     return list(filter(lambda payment: payment["iban"] == iban, payments))
 
 
 def get_exchange_rates():
     """Vrátí kurzy uložené v souboru (cache)."""
-    with open(database_path / "exchange_rates.yaml", "r", encoding="utf8") as file:
+    with open(EXCHANGE_RATES_PATH, "r", encoding="utf8") as file:
         rates_dictionary = yaml.safe_load(file)
     return rates_dictionary
 
 
 def set_exchange_rates(rates_dictionary):
     """Uloží nové kurzy do souboru."""
-    with open(database_path / "exchange_rates.yaml", "w", encoding="utf8") as file:
+    with open(EXCHANGE_RATES_PATH, "w", encoding="utf8") as file:
         yaml.dump(rates_dictionary, file)
